% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3cache.R
\name{exists}
\alias{exists}
\title{Check for the existence of a value in the s3cache at a particular key.}
\usage{
exists(key, path = default_path())
}
\arguments{
\item{key}{character. The key to store at.}

\item{path}{character. Path to the s3bucket if you want to define a new cache location.}
}
\description{
Check for the existence of a value in the s3cache at a particular key.
}

