% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3cache.R
\name{forget}
\alias{forget}
\title{Delete the s3cache at a particular key.}
\usage{
forget(key, path = default_path())
}
\arguments{
\item{key}{character. The key to store at.}

\item{path}{character. Path to the s3bucket if you want to define a new cache location.}
}
\description{
Delete the s3cache at a particular key.
}

