% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3cache.R
\name{set}
\alias{set}
\title{Store a value in the s3cache at a particular key.}
\usage{
set(key, value, path = default_path(), expires_in = NULL)
}
\arguments{
\item{key}{character. The key to store at.}

\item{value}{character. The value to store.}

\item{path}{character. Path to the s3bucket if you want to define a new cache location.}

\item{expires_in}{character. A string defining when the cache expires or the time of expiration (POSIXt).}
}
\description{
Store a value in the s3cache at a particular key.
}

